fixed javascript to take a timing arg in all cases

+++ javascript/run.sh
@@ -1,14 +1,28 @@
 #!/bin/sh
 INTERP="$1"
-shift
+COUNT="$2"
+shift 2
+if [ $# -ne 0 ]
+then
+    echo "usage" >&2
+    exit 1
+fi
+
+TMP=/tmp/js.$$
+trap "rm -f $TMP" 0 1 2 3 15
+cat gamevalue.js negamax.js ttt.js >$TMP
+
 case $INTERP in
 d8)
-    d8 "$@" gamevalue.js negamax.js ttt.js
+    d8 $TMP -- $COUNT
     ;;
-js|rhino)
-    TMP=/tmp/js.$$
-    trap "rm -f $TMP" 0 1 2 3 15
-    cat gamevalue.js negamax.js ttt.js >$TMP
-    $INTERP "$@" $TMP
+smjs)
+    smjs $TMP -- $COUNT
     ;;
+rhino)
+    rhino $TMP $COUNT
+    ;;
+*)
+    echo "unknown interpreter" >&2
+    exit 1
 esac
